axum: Move `OriginalUri` to dedicated module
This seems unrelated to "request parts", so let's give it a more reasonable name.
